+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 * [#3615](https://github.com/bbatsov/rubocop/pull/3615): Add autocorrection for `Lint/EmptyInterpolation`. ([@pocke][])
 * Make `PercentLiteralDelimiters` enforce delimiters around `%I()` too. ([@bronson][])
+* [#3408](https://github.com/bbatsov/rubocop/issues/3408): Add check for repeated values in case conditionals. ([@swcraig][])
 
 ### Bug fixes
 
@@ -2445,3 +2446,4 @@
 [@bronson]: https://github.com/bronson
 [@albus522]: https://github.com/albus522
 [@sihu]: https://github.com/sihu
+[@swcraig]: https://github.com/swcraig